+++ lldb/source/API/SBThread.cpp
@@ -1331,8 +1331,6 @@ bool SBThread::SafeToCallFunctions() {
   return true;
 }
 
-lldb::ThreadSP SBThread::GetSP() const { return m_opaque_sp->GetThreadSP(); }
-
 lldb_private::Thread *SBThread::operator->() {
   return get();
 }