+++ clang/docs/ReleaseNotes.rst
@@ -328,6 +328,7 @@ Python Binding Changes
 
 OpenMP Support
 --------------
+- Added support for 'omp assume' directive.
 
 Improvements
 ^^^^^^^^^^^^